Move `pidfile` into its own file, as it has grown substantially in size
